Refactor: 부스 예약 서비스 분리
