+++ src/main/java/com/openbook/openbook/booth/controller/UserBoothController.java
@@ -4,6 +4,7 @@ import com.openbook.openbook.booth.controller.request.BoothRegistrationRequest;
 import com.openbook.openbook.booth.controller.response.*;
 import com.openbook.openbook.booth.service.BoothCommonService;
 import com.openbook.openbook.booth.service.common.CommonProductService;
+import com.openbook.openbook.booth.service.common.CommonReservationService;
 import com.openbook.openbook.booth.service.core.BoothReservationService;
 import com.openbook.openbook.global.dto.ResponseMessage;
 import com.openbook.openbook.global.dto.SliceResponse;
@@ -25,7 +26,7 @@ public class UserBoothController {
 
     private final BoothCommonService boothCommonService;
     private final CommonProductService commonProductService;
-    private final BoothReservationService boothReservationService;
+    private final CommonReservationService commonReservationService;
 
     @PostMapping
     public ResponseEntity <ResponseMessage>  registration(Authentication authentication, @Valid BoothRegistrationRequest request){
@@ -84,7 +85,7 @@ public class UserBoothController {
 
     @GetMapping("/{booth_id}/reservations")
     public ResponseEntity<List<BoothReservationsResponse>> getAllBoothReservations(@PathVariable Long booth_id){
-        return ResponseEntity.ok(boothReservationService.getAllBoothReservations(booth_id));
+        return ResponseEntity.ok(commonReservationService.getAllBoothReservations(booth_id));
     }
 
 }
